Fixed search bar bug
Given a user is using the search functionality,
When a weather lookup is completed using current location,
Then all elements of the page will reflect the current location.

A bug was identified when using current location after searching by city name, the value of the search bar retained the previous search, rather than reflecting the current location.
This has been resolved by pulling the current location name in to the search bar when using the 📍 button.